Creates SafeLazy to be used by all caching things instead of relying on the base class as a static that cannot be moved right now.
