+++ src/Umbraco.Core/Cache/FastDictionaryAppCacheBase.cs
@@ -2,7 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.ExceptionServices;
 using System.Text.RegularExpressions;
 using Umbraco.Core.Composing;
 
@@ -15,9 +14,6 @@ namespace Umbraco.Core.Cache
     {
         // prefix cache keys so we know which one are ours
         protected const string CacheItemPrefix = "umbrtmche";
-
-        // an object that represent a value that has not been created yet
-        protected internal static readonly object ValueNotCreated = new object();
 
         #region IAppCache
 
@@ -35,7 +31,7 @@ namespace Umbraco.Core.Cache
             {
                 ExitReadLock();
             }
-            return result == null ? null : GetSafeLazyValue(result); // return exceptions as null
+            return result == null ? null : SafeLazy.GetSafeLazyValue(result); // return exceptions as null
         }
 
         /// <inheritdoc />
@@ -59,7 +55,7 @@ namespace Umbraco.Core.Cache
             }
 
             return entries
-                .Select(x => GetSafeLazyValue((Lazy<object>)x.Value)) // return exceptions as null
+                .Select(x => SafeLazy.GetSafeLazyValue((Lazy<object>)x.Value)) // return exceptions as null
                 .Where(x => x != null); // backward compat, don't store null values in the cache
         }
 
@@ -82,7 +78,7 @@ namespace Umbraco.Core.Cache
                 ExitReadLock();
             }
             return entries
-                .Select(x => GetSafeLazyValue((Lazy<object>)x.Value)) // return exceptions as null
+                .Select(x => SafeLazy.GetSafeLazyValue((Lazy<object>)x.Value)) // return exceptions as null
                 .Where(x => x != null); // backward compatible, don't store null values in the cache
         }
 
@@ -132,7 +128,7 @@ namespace Umbraco.Core.Cache
                         // entry.Value is Lazy<object> and not null, its value may be null
                         // remove null values as well, does not hurt
                         // get non-created as NonCreatedValue & exceptions as null
-                        var value = GetSafeLazyValue((Lazy<object>) x.Value, true);
+                        var value = SafeLazy.GetSafeLazyValue((Lazy<object>) x.Value, true);
 
                         // if T is an interface remove anything that implements that interface
                         // otherwise remove exact types (not inherited types)
@@ -162,7 +158,7 @@ namespace Umbraco.Core.Cache
                         // remove null values as well, does not hurt
                         // compare on exact type, don't use "is"
                         // get non-created as NonCreatedValue & exceptions as null
-                        var value = GetSafeLazyValue((Lazy<object>) x.Value, true);
+                        var value = SafeLazy.GetSafeLazyValue((Lazy<object>) x.Value, true);
 
                         // if T is an interface remove anything that implements that interface
                         // otherwise remove exact types (not inherited types)
@@ -193,7 +189,7 @@ namespace Umbraco.Core.Cache
                         // remove null values as well, does not hurt
                         // compare on exact type, don't use "is"
                         // get non-created as NonCreatedValue & exceptions as null
-                        var value = GetSafeLazyValue((Lazy<object>) x.Value, true);
+                        var value = SafeLazy.GetSafeLazyValue((Lazy<object>) x.Value, true);
                         if (value == null) return true;
 
                         // if T is an interface remove anything that implements that interface
@@ -272,57 +268,7 @@ namespace Umbraco.Core.Cache
             return $"{CacheItemPrefix}-{key}";
         }
 
-        protected internal static Lazy<object> GetSafeLazy(Func<object> getCacheItem)
-        {
-            // try to generate the value and if it fails,
-            // wrap in an ExceptionHolder - would be much simpler
-            // to just use lazy.IsValueFaulted alas that field is
-            // internal
-            return new Lazy<object>(() =>
-            {
-                try
-                {
-                    return getCacheItem();
-                }
-                catch (Exception e)
-                {
-                    return new ExceptionHolder(ExceptionDispatchInfo.Capture(e));
-                }
-            });
-        }
-
-        protected internal static object GetSafeLazyValue(Lazy<object> lazy, bool onlyIfValueIsCreated = false)
-        {
-            // if onlyIfValueIsCreated, do not trigger value creation
-            // must return something, though, to differentiate from null values
-            if (onlyIfValueIsCreated && lazy.IsValueCreated == false) return ValueNotCreated;
-
-            // if execution has thrown then lazy.IsValueCreated is false
-            // and lazy.IsValueFaulted is true (but internal) so we use our
-            // own exception holder (see Lazy<T> source code) to return null
-            if (lazy.Value is ExceptionHolder) return null;
-
-            // we have a value and execution has not thrown so returning
-            // here does not throw - unless we're re-entering, take care of it
-            try
-            {
-                return lazy.Value;
-            }
-            catch (InvalidOperationException e)
-            {
-                throw new InvalidOperationException("The method that computes a value for the cache has tried to read that value from the cache.", e);
-            }
-        }
-
-        internal class ExceptionHolder
-        {
-            public ExceptionHolder(ExceptionDispatchInfo e)
-            {
-                Exception = e;
-            }
-
-            public ExceptionDispatchInfo Exception { get; }
-        }
+       
 
         #endregion
     }
