+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -124,6 +124,8 @@
       <DependentUpon>Constants.cs</DependentUpon>
     </Compile>
     -->
+    <Compile Include="Cache\FastDictionaryAppCacheBase.cs" />
+    <Compile Include="Cache\ObjectCacheAppCache.cs" />
     <Compile Include="Compose\AuditEventsComponent.cs" />
     <Compile Include="Cache\AppCaches.cs" />
     <Compile Include="Cache\CacheRefresherBase.cs" />
@@ -133,12 +135,10 @@
     <Compile Include="Cache\DeepCloneAppCache.cs" />
     <Compile Include="Cache\DefaultRepositoryCachePolicy.cs" />
     <Compile Include="Cache\FastDictionaryAppCache.cs" />
-    <Compile Include="Cache\FastDictionaryAppCacheBase.cs" />
     <Compile Include="Cache\FullDataSetRepositoryCachePolicy.cs" />
     <Compile Include="Cache\HttpRequestAppCache.cs" />
     <Compile Include="Cache\WebCachingAppCache.cs" />
     <Compile Include="Cache\JsonCacheRefresherBase.cs" />
-    <Compile Include="Cache\ObjectCacheAppCache.cs" />
     <Compile Include="Cache\PayloadCacheRefresherBase.cs" />
     <Compile Include="Cache\RepositoryCachePolicyBase.cs" />
     <Compile Include="Cache\RepositoryCachePolicyOptions.cs" />
